[O2B-416] Change filePreview functionality to a separate directory
