+++ lib/public/views/Logs/Create/LogCreationModel.js
@@ -15,6 +15,7 @@ import { jsonFetch } from '../../../utilities/fetch/jsonFetch.js';
 import { tagToOption } from '../../../components/tag/tagToOption.js';
 import { getRemoteData } from '../../../utilities/fetch/getRemoteData.js';
 import { TagPickerModel } from '../../../components/tag/TagPickerModel.js';
+import { FilePreviewModel } from '../FilePreview/FilePreviewModel.js'
 
 const DATA_DISPLAY_DELIMITER = ',';
 
@@ -43,6 +44,9 @@ export class LogCreationModel extends Observable {
         this._creationTagsPickerModel.bubbleTo(this);
         this._creationTagsPickerModel.visualChange$.bubbleTo(this);
 
+        this._filePreviewModel = new FilePreviewModel();
+        this._filePreviewModel.bubbleTo(this)
+
         this._runNumbers = runNumbers.join(DATA_DISPLAY_DELIMITER);
         this._lhcFills = lhcFillNumbers.join(DATA_DISPLAY_DELIMITER);
         this._environments = environmentIds.join(DATA_DISPLAY_DELIMITER);
@@ -69,7 +73,6 @@ export class LogCreationModel extends Observable {
         this._isRunNumbersReadonly = this._runNumbers;
 
         this._textEditor = null;
-        this._attachmentPreviews = {};
     }
 
     /**
@@ -412,27 +415,7 @@ export class LogCreationModel extends Observable {
         this._textEditor = value;
     }
 
-    /**
-     * Returns the data for the file preview
-     *
-     * @param {string} fileName the name of the file you want a preview of
-     * @returns {string|vnode|object} the data for the preview
-     */
-    getAttachmentPreview(fileName) {
-        return this._attachmentPreviews[fileName];
-    }
-
-    /**
-     * Stores the new file preview data if the file name does not already have preview data
-     *
-     * @param {string|vnode|object} preview the file preview data
-     * @param {string} fileName the name of the file for the preview
-     * @returns {void}
-     */
-    setAttachmentPreview(preview, fileName) {
-        if (!this._attachmentPreviews[fileName]) {
-            this._attachmentPreviews[fileName] = preview;
-            this.notify();
-        }
+    get filePreviewModel() {
+        return this._filePreviewModel;
     }
 }
