+++ lib/public/views/Logs/Create/logCreationComponent.js
@@ -18,9 +18,7 @@ import { tagPicker } from '../../../components/tag/tagPicker.js';
 import { markdownInput } from '../../../components/common/markdown/markdown.js';
 import spinner from '../../../components/common/spinner.js';
 import { logComponent } from '../../../components/Log/log.js';
-import { popover } from '../../../components/common/popover/popover.js';
-import { PopoverAnchors } from '../../../components/common/popover/PopoverEngine.js';
-import { PopoverTriggerPreConfiguration } from '../../../components/common/popover/popoverPreConfigurations.js';
+import { showFilePreview } from '../FilePreview/filePreviewComponent.js';
 
 /**
  * A function to construct the create log screen
@@ -239,7 +237,20 @@ const environmentsPanel = (logCreationModel) => labeledInput(
  * @return {vnode} - panel allowing the user to attach files
  */
 const attachmentsPanel = (logCreationModel) => {
-    const { attachments } = logCreationModel;
+    const { attachments, filePreviewModel } = logCreationModel;
+
+    let fileText;
+    switch(attachments.length) {
+        case 0:
+            fileText = 'No files chosen';
+            break;
+        case 1:
+            fileText = `${attachments.length} file chosen`;
+            break;
+        default:
+            fileText = `${attachments.length} files chosen`;
+            break;
+    }
 
     return labeledInput(
         { for: 'attachments' },
@@ -251,11 +262,7 @@ const attachmentsPanel = (logCreationModel) => {
                     style: 'display: flex; align-items: center; width: -webkit-fill-available;',
                 }, [
                     h('.btn', 'Choose files'),
-                    h('', { style: 'margin-left: 1%;' }, [
-                        attachments.length === 0 && h('', 'No files chosen'),
-                        attachments.length === 1 && h('', `${attachments.length} file chosen`),
-                        attachments.length > 1 && h('', `${attachments.length} files chosen`),
-                    ]),
+                    h('', { style: 'margin-left: 1%;' }, fileText),
                 ]),
                 h('input#attachments.w-33', {
                     type: 'file',
@@ -271,167 +278,9 @@ const attachmentsPanel = (logCreationModel) => {
             ]),
             attachments.length > 0 && h('#attachmentNames', {
                 style: 'min-height: 1.5em;',
-            }, [...attachments].map((attachment) => [showFilePreview(attachment, logCreationModel)])),
+            }, [...attachments].map((attachment) => [showFilePreview(filePreviewModel, attachment)])),
         ]),
     );
-};
-
-/**
- * Creates a preview of the file given by returning a popover. If creating a preview is not possible,
- * just a vnode for the file name is returned.
- * The following file extensions are tested and implemented, not tested and implemented, or not possible:
- *      Tested and implemented:
- *      - png
- *      - jpg
- *      - txt
- *      - log
- *      - sh
- *      - c
- *      - jpeg
- *      - json
- *      - csv
- *      - gif (image based)
- *
- *      Not tested or implemented as preview:
- *      - pages
- *      - zip
- *      - rtf
- *      - gz
- *      - view1
- *
- *      Not possible (since you need an external url or library):
- *      - docx
- *      - odt
- *      - pdf
- *      - xlsx
- *      - pptx
- *
- * @param {File} attachment the file to create a preview of, if possible
- * @param {LogCreationModel} logCreationModel logs creation model for getting and setting the attachment previews
- * @returns {vnode} the file name with or without a popover
- */
-const showFilePreview = (attachment, logCreationModel) => {
-    const previewFileTypes = {
-        image: false,
-        json: false,
-        csv: false,
-        shellscript: false,
-        text: false, // For all text file types left
-    };
-
-    for (const key in previewFileTypes) {
-        if (attachment.type.match(key)) {
-            previewFileTypes[key] = true;
-        }
-    }
-
-    /**
-     * Creates a vnode with the file name and a line break at the end.
-     *
-     * @param {string} style the style to apply, if any
-     * @returns {vnode} the vnode with the file name and a line break
-     */
-    const fileName = (style = '') => h('span.attachment', {
-        key: attachment.name,
-        style,
-    }, [attachment.name, h('br')]);
-
-    /**
-     * Creates a vnode with a clickable file name to open the popover with the preview
-     *
-     * @param {string|vnode} content the preview to add to the popover
-     * @returns {vnode} the preview popover
-     */
-    const previewPopover = (content) => popover(
-        fileName('text-decoration: underline; cursor: pointer;'),
-        h('', {
-            style: 'margin: 1%; max-height: 99%; overflow: hidden; display: flex; flex-direction: column;',
-        }, [
-            h('.f3', {}, attachment.name),
-            h('', {
-                style: 'overflow: auto;',
-            }, content),
-        ]),
-        {
-            ...PopoverTriggerPreConfiguration.click,
-            anchor: PopoverAnchors.LEFT_MIDDLE,
-        },
-    );
-
-    const previewFileStyle = 'margin: 0; width: 100%;';
-
-    if (previewFileTypes.image) {
-        const reader = new FileReader();
-        const fileContent = h('img', {
-            alt: attachment.name,
-            style: previewFileStyle,
-            src: logCreationModel.getAttachmentPreview(attachment.name),
-        });
-
-        reader.onload = () => {
-            logCreationModel.setAttachmentPreview(reader.result, attachment.name);
-        };
-        reader.readAsDataURL(attachment);
-
-        return previewPopover(fileContent);
-    } else if (previewFileTypes.csv) {
-        const reader = new FileReader();
-        const fileContent = h('table.table.table-hover', {
-            style: previewFileStyle,
-        }, logCreationModel.getAttachmentPreview(attachment.name));
-
-        reader.onload = () => {
-            let content;
-
-            const lines = reader.result.split('\n');
-            if (lines.length < 2) {
-                content = h('div', 'Invalid CSV format');
-            } else {
-                const headers = lines[0].split(',');
-                const headerRow = h('tr', headers.map((header, index) => h('th', {
-                    key: index,
-                }, header)));
-
-                const tableRows = lines.slice(1).map((line, rowIndex) => {
-                    const columns = line.split(',');
-                    return h('tr', { key: rowIndex }, columns.map((column, colIndex) => h('td', { key: colIndex }, column)));
-                });
-
-                content = [
-                    h('thead', { key: 'thead' }, headerRow),
-                    h('tbody', { key: 'tbody' }, tableRows),
-                ];
-            }
-
-            logCreationModel.setAttachmentPreview(content, attachment.name);
-        };
-        reader.readAsText(attachment);
-
-        return previewPopover(fileContent);
-    } else if (previewFileTypes.json || previewFileTypes.text || previewFileTypes.shellscript) {
-        const reader = new FileReader();
-        const fileContent = h('pre', {
-            style: previewFileStyle,
-        }, logCreationModel.getAttachmentPreview(attachment.name));
-
-        reader.onload = () => {
-            let content;
-
-            if (previewFileTypes.json) {
-                const jsonContent = JSON.parse(reader.result);
-                content = JSON.stringify(jsonContent, null, 2);
-            } else if (previewFileTypes.text || previewFileTypes.shellscript) {
-                content = reader.result;
-            }
-
-            logCreationModel.setAttachmentPreview(content, attachment.name);
-        };
-        reader.readAsText(attachment);
-
-        return previewPopover(fileContent);
-    } else {
-        return fileName();
-    }
 };
 
 /**
